Battery App: PMIC and pre v5 (none pmic) now use the same battery app.
