+++ app/boards/arm/zswatch_nrf5340/zswatch_nrf5340_cpuapp_5.conf
@@ -43,3 +43,7 @@ CONFIG_APPLICATIONS_USE_BATTERY=n
 # Temporary due to current flash having larger pages than NVS can handle.
 CONFIG_FCB=y
 CONFIG_SETTINGS_FCB=y
+CONFIG_SETTINGS_NVS=n
+CONFIG_NVS=n
+
+CONFIG_ERASE_PROGRESSIVELY=n
