+++ app/prj.conf
@@ -115,7 +115,6 @@ CONFIG_FLASH_MAP_LABELS=y
 
 CONFIG_SETTINGS=y
 CONFIG_NVS=y
-CONFIG_SETTINGS_NVS=y
 
 # Bluetooth
 CONFIG_BT=y
