+++ app/src/applications/battery/CMakeLists.txt
@@ -1,4 +1,2 @@
-if(CONFIG_APPLICATIONS_USE_BATTERY)
-    FILE(GLOB app_sources *.c)
-    target_sources(app PRIVATE ${app_sources})
-endif()
+FILE(GLOB app_sources *.c)
+target_sources(app PRIVATE ${app_sources})
