+++ app/src/applications/battery/battery_app.c
@@ -2,29 +2,47 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/init.h>
 #include <zephyr/zbus/zbus.h>
+#include <zephyr/settings/settings.h>
 
 #include "battery/battery_ui.h"
 #include "events/battery_event.h"
 #include "managers/zsw_app_manager.h"
 #include "ui/utils/zsw_ui_utils.h"
+#include "fuel_gauge/zsw_pmic.h"
+#include "battery_ui.h"
 
-LOG_MODULE_REGISTER(battery_app, LOG_LEVEL_WRN);
+LOG_MODULE_REGISTER(pmic_app, LOG_LEVEL_WRN);
+
+#define SETTING_BATTERY_HIST "battery/hist"
+
+#define MAX_SAMPLES (7 * 24 * 4) // 7 days of 15 minute samples
+#define SAMPLE_INTERVAL (15 * 60 * 1000) // 15 minutes in milliseconds
 
 static void battery_app_start(lv_obj_t *root, lv_group_t *group);
 static void battery_app_stop(void);
 
 static void zbus_battery_sample_data_callback(const struct zbus_channel *chan);
-static int get_num_samples(void);
+static void on_battery_hist_clear_cb(void);
 
 typedef struct battery_sample_t {
-    int mV;
-    int64_t timestamp;
+    uint8_t mv_with_decimals; // Calced by: 3.65V - 2 => 1.65 => 165
+    uint8_t percent;
 } battery_sample_t;
+
+typedef struct battery_storage_t {
+    uint16_t first_sample_index;
+    uint16_t num_samples;
+    battery_sample_t samples[MAX_SAMPLES];
+} battery_storage_t;
 
 ZBUS_CHAN_DECLARE(battery_sample_data_chan);
 ZBUS_LISTENER_DEFINE(battery_app_battery_event, zbus_battery_sample_data_callback);
+ZBUS_CHAN_ADD_OBS(battery_sample_data_chan, battery_app_battery_event, 1);
 
 LV_IMG_DECLARE(battery_app_icon);
+
+static battery_storage_t battery_context = {0};
+static uint64_t last_battery_sample_time = 0;
 
 static application_t app = {
     .name = "Battery",
@@ -33,26 +51,24 @@ static application_t app = {
     .stop_func = battery_app_stop
 };
 
-static battery_sample_t battery_samples[CONFIG_DEFAULT_CONFIGURATION_BATTERY_NUM_SAMPLES_MAX];
-static int next_battery_sample_index;
-
 static void battery_app_start(lv_obj_t *root, lv_group_t *group)
 {
-    struct battery_sample_event last_sample;
-
-    if (!zbus_chan_read(&battery_sample_data_chan, &last_sample, K_MSEC(100)) == 0) {
-        LOG_ERR("Failed to read battery sample data");
-        return;
+    battery_sample_t *sample;
+    struct battery_sample_event initial_sample;
+#if CONFIG_DT_HAS_NORDIC_NPM1300_ENABLED
+    battery_ui_show(root, on_battery_hist_clear_cb, battery_context.num_samples + 1, true);
+#else
+    battery_ui_show(root, on_battery_hist_clear_cb, battery_context.num_samples + 1, false);
+#endif
+    for (int i = 0; i < battery_context.num_samples; i++) {
+        sample = &battery_context.samples[(battery_context.first_sample_index + i) % MAX_SAMPLES];
+        battery_ui_add_measurement(sample->percent, (sample->mv_with_decimals * 10) + 2000);
     }
 
-    battery_ui_show(root, get_num_samples() + 1);
-    battery_ui_set_current_measurement(last_sample.mV);
-    for (int i = 0; i < CONFIG_DEFAULT_CONFIGURATION_BATTERY_NUM_SAMPLES_MAX; i++) {
-        if (battery_samples[(next_battery_sample_index + i) % CONFIG_DEFAULT_CONFIGURATION_BATTERY_NUM_SAMPLES_MAX].timestamp !=
-            0) {
-            battery_ui_add_measurement(battery_samples[(next_battery_sample_index + i) %
-                                                       CONFIG_DEFAULT_CONFIGURATION_BATTERY_NUM_SAMPLES_MAX].mV);
-        }
+    if (zbus_chan_read(&battery_sample_data_chan, &initial_sample, K_MSEC(100)) == 0) {
+        battery_ui_update(initial_sample.ttf, initial_sample.tte, initial_sample.status, initial_sample.error,
+                       initial_sample.is_charging);
+        battery_ui_add_measurement(initial_sample.percent, initial_sample.mV);
     }
 }
 
@@ -61,46 +77,89 @@ static void battery_app_stop(void)
     battery_ui_remove();
 }
 
+static void add_battery_sample(const struct battery_sample_event *event)
+{
+    int next_battery_sample_index = (battery_context.first_sample_index + battery_context.num_samples) % MAX_SAMPLES;
+    battery_context.samples[next_battery_sample_index].mv_with_decimals = ((event->mV - 2000) / 1000) * 100 + ((
+                                                                              event->mV / 10) % 100);
+    battery_context.samples[next_battery_sample_index].percent = event->percent;
+
+    battery_context.num_samples++;
+    if (battery_context.num_samples > MAX_SAMPLES) {
+        battery_context.first_sample_index = (battery_context.first_sample_index + 1) % MAX_SAMPLES;
+        battery_context.num_samples = MAX_SAMPLES;
+    }
+    LOG_DBG("%d, %d, %d", battery_context.first_sample_index, battery_context.num_samples, next_battery_sample_index);
+}
+
 static void zbus_battery_sample_data_callback(const struct zbus_channel *chan)
 {
     const struct battery_sample_event *event = zbus_chan_const_msg(chan);
-    int previous_sample_index;
 
-    if (next_battery_sample_index == 0) {
-        previous_sample_index = CONFIG_DEFAULT_CONFIGURATION_BATTERY_NUM_SAMPLES_MAX - 1;
-    } else {
-        previous_sample_index = next_battery_sample_index - 1;
+    if ((k_uptime_get() - last_battery_sample_time) >= SAMPLE_INTERVAL) {
+        add_battery_sample(event);
+        int ret = settings_save_one(SETTING_BATTERY_HIST, &battery_context, sizeof(battery_storage_t));
+        if (ret) {
+            LOG_ERR("Error during saving of battery_context.samples %d", ret);
+        }
+        last_battery_sample_time = k_uptime_get();
+        battery_ui_add_measurement(event->percent, event->mV);
     }
-    if (battery_samples[previous_sample_index].timestamp == 0 ||
-        ((k_uptime_get() - battery_samples[previous_sample_index].timestamp) >=
-         (CONFIG_DEFAULT_CONFIGURATION_BATTERY_SAMPLE_INTERVAL_MINUTES * 60 * 1000))) {
-        battery_samples[next_battery_sample_index].mV = event->mV;
-        battery_samples[next_battery_sample_index].timestamp = k_uptime_get();
-        LOG_DBG("Add %d\n", battery_samples[next_battery_sample_index].mV);
-        next_battery_sample_index = (next_battery_sample_index + 1) % CONFIG_DEFAULT_CONFIGURATION_BATTERY_NUM_SAMPLES_MAX;
-    } else {
-        LOG_DBG("Discard sample: %d, %d\n", battery_samples[previous_sample_index].mV,
-                (int)(k_uptime_get() - battery_samples[previous_sample_index].timestamp));
+    battery_ui_update(event->ttf, event->tte, event->status, event->error, event->is_charging);
+}
+
+static void on_battery_hist_clear_cb(void)
+{
+    memset(battery_context.samples, 0, sizeof(battery_context.samples));
+    if (settings_delete(SETTING_BATTERY_HIST) != 0) {
+        LOG_ERR("Error during settings_delete!");
     }
 }
 
-static int get_num_samples(void)
+static int battery_load_state(const char *p_key, size_t len,
+                              settings_read_cb read_cb, void *p_cb_arg, void *p_param)
 {
-    int num_samples = 0;
+    ARG_UNUSED(p_key);
 
-    for (int i = 0; i < CONFIG_DEFAULT_CONFIGURATION_BATTERY_NUM_SAMPLES_MAX; i++) {
-        if (battery_samples[i].timestamp != 0) {
-            num_samples++;
-        }
+    if (len != sizeof(battery_storage_t)) {
+        return -EINVAL;
     }
 
-    return num_samples;
+    int num_bytes_read = read_cb(p_cb_arg, &battery_context, len);
+    LOG_DBG("Read %d bytes", num_bytes_read);
+
+    LOG_DBG("Number samples: %d", battery_context.num_samples);
+
+    if (num_bytes_read == 0) {
+        LOG_ERR("Currupt battery settings data");
+    } else if (num_bytes_read != sizeof(battery_storage_t)) {
+        LOG_WRN("No battery settings data read");
+        return -ENODATA;
+    }
+
+    LOG_DBG("Loaded battery history data");
+
+    return 0;
 }
 
 static int battery_app_add(void)
 {
     zsw_app_manager_add_application(&app);
-    memset(battery_samples, 0, sizeof(battery_samples));
+
+    if (settings_subsys_init()) {
+        LOG_ERR("Error during settings_subsys_init!");
+        return -EFAULT;
+    }
+
+    if (settings_load_subtree_direct(SETTING_BATTERY_HIST, battery_load_state, NULL)) {
+        LOG_ERR("Error during settings_load_subtree!");
+        return -EFAULT;
+    }
+
+    if (settings_delete(SETTING_BATTERY_HIST) != 0) {
+        LOG_ERR("Error during settings_delete!");
+    }
+
     return 0;
 }
